backend/requirements.txt
fastapi==0.116.1
google-cloud-firestore==2.21.0
uvicorn==0.35.0
